refactor(courtboard): extract panel components
- Extract NextAvailablePanel to components/NextAvailablePanel.jsx
- Extract ReservedCourtsPanel to components/ReservedCourtsPanel.jsx
- Move selectReservedItemsFromBlocks and normalizeBlock helpers
  to ReservedCourtsPanel.jsx
- Pass maxDisplay as prop instead of using config directly
- Remove inline definitions from main.jsx (~412 lines removed)
